Refactor total utils and Helpers
