+++ src/me/corriekay/pokegoutil/controllers/AccountControllerNew.java
@@ -5,12 +5,8 @@ import com.pokegoapi.api.player.PlayerProfile;
 import com.pokegoapi.auth.CredentialProvider;
 import com.pokegoapi.auth.GoogleUserCredentialProvider;
 import com.pokegoapi.auth.PtcCredentialProvider;
-import com.pokegoapi.exceptions.LoginFailedException;
-import com.pokegoapi.exceptions.RemoteServerException;
-import me.corriekay.pokegoutil.utils.Browser;
+import me.corriekay.pokegoutil.utils.helpers.Browser;
 import me.corriekay.pokegoutil.utils.Config;
-import me.corriekay.pokegoutil.utils.Console;
-import me.corriekay.pokegoutil.windows.PokemonGoMainWindow;
 import okhttp3.OkHttpClient;
 
 import javax.swing.*;
