--- src/me/corriekay/pokegoutil/utils/LDocumentListener.java
@@ -1,67 +1,0 @@
-package me.corriekay.pokegoutil.utils;
-
-import java.beans.PropertyChangeEvent;
-import java.beans.PropertyChangeListener;
-import java.util.Objects;
-
-import javax.swing.*;
-import javax.swing.event.*;
-import javax.swing.text.Document;
-import javax.swing.text.JTextComponent;
-
-public class LDocumentListener {
-	/**
-	 * Installs a listener to receive notification when the text of any
-	 * {@code JTextComponent} is changed. Internally, it installs a
-	 * {@link DocumentListener} on the text component's {@link Document},
-	 * and a {@link PropertyChangeListener} on the text component to detect
-	 * if the {@code Document} itself is replaced.
-	 * 
-	 * @param text any text component, such as a {@link JTextField}
-	 *        or {@link JTextArea}
-	 * @param changeListener a listener to receieve {@link ChangeEvent}s
-	 *        when the text is changed; the source object for the events
-	 *        will be the text component
-	 * @throws NullPointerException if either parameter is null
-	 * @author http://stackoverflow.com/users/964243/boann
-	 * @source http://stackoverflow.com/a/27190162/1340219
-	 */
-	public static void addChangeListener(JTextComponent text, ChangeListener changeListener) {
-	    Objects.requireNonNull(text);
-	    Objects.requireNonNull(changeListener);
-	    DocumentListener dl = new DocumentListener() {
-	        private int lastChange = 0, lastNotifiedChange = 0;
-	
-	        @Override
-	        public void insertUpdate(DocumentEvent e) {
-	            changedUpdate(e);
-	        }
-	
-	        @Override
-	        public void removeUpdate(DocumentEvent e) {
-	            changedUpdate(e);
-	        }
-	
-	        @Override
-	        public void changedUpdate(DocumentEvent e) {
-	            lastChange++;
-	            SwingUtilities.invokeLater(() -> {
-	                if (lastNotifiedChange != lastChange) {
-	                    lastNotifiedChange = lastChange;
-	                    changeListener.stateChanged(new ChangeEvent(text));
-	                }
-	            });
-	        }
-	    };
-	    text.addPropertyChangeListener("document", (PropertyChangeEvent e) -> {
-	        Document d1 = (Document)e.getOldValue();
-	        Document d2 = (Document)e.getNewValue();
-	        if (d1 != null) d1.removeDocumentListener(dl);
-	        if (d2 != null) d2.addDocumentListener(dl);
-	        dl.changedUpdate(null);
-	    });
-	    Document d = text.getDocument();
-	    if (d != null) d.addDocumentListener(dl);
-	}
-}
-
