+++ src/me/corriekay/pokegoutil/windows/PokemonTab.java
@@ -11,7 +11,12 @@ import com.pokegoapi.api.map.pokemon.EvolutionResult;
 import com.pokegoapi.api.player.PlayerProfile.Currency;
 import com.pokegoapi.api.pokemon.*;
 import me.corriekay.pokegoutil.utils.*;
-import me.corriekay.pokegoutil.utils.PokemonCpUtils;
+import me.corriekay.pokegoutil.utils.helpers.JTableColumnPacker;
+import me.corriekay.pokegoutil.utils.helpers.LDocumentListener;
+import me.corriekay.pokegoutil.utils.pokemon.PokeHandler;
+import me.corriekay.pokegoutil.utils.pokemon.PokemonCpUtils;
+import me.corriekay.pokegoutil.utils.helpers.DateHelper;
+import me.corriekay.pokegoutil.utils.ui.GhostText;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.math.NumberUtils;
 import org.apache.commons.lang3.mutable.MutableInt;
